Move /pages/{Home -> index}.js
